ValueNodeRegistry improvements
- move into separate files
- get rid of unused canvas value in create() method
- add macros for registrable node definitions
