+++ synfig-core/src/synfig/canvas.cpp
@@ -47,6 +47,7 @@
 #include "importer.h"
 #include "layer.h"
 #include "loadcanvas.h"
+#include "valuenode_registry.h"
 
 #include "layers/layer_pastecanvas.h"
 #include "valuenodes/valuenode_const.h"
@@ -1404,7 +1405,7 @@ synfig::optimize_layers(Time time, Context context, Canvas::Handle op_canvas, bo
 				amount=ValueNode_Const::create(layer->get_param("amount").get(Real()));
 			}
 			// Connect a ValueNode_Scale to the amount parameter with the right sub-parameters
-			ValueNode::Handle value_node=ValueNodeRegistry::create("scale", ValueBase(Real()), op_canvas);
+			ValueNode::Handle value_node=ValueNodeRegistry::create("scale", ValueBase(Real()));
 			ValueNode_Scale::Handle scale=ValueNode_Scale::Handle::cast_dynamic(value_node);
 			scale->set_link("link", amount);
 			scale->set_link("scalar", ValueNode_Const::create(layer_visibility));
