+++ synfig-core/src/synfig/valuenode.h
@@ -32,16 +32,17 @@
 #include "value.h"
 #include "string.h"
 #include "releases.h"
-#include <ETL/handle>
-#include <ETL/stringf>
 #include "exception.h"
-#include <sigc++/signal.h>
 #include "guid.h"
-#include <ETL/angle>
 #include "paramdesc.h"
 #include "interpolation.h"
-
 #include "node.h"
+
+#include <ETL/angle>
+#include <ETL/handle>
+#include <ETL/stringf>
+
+#include <sigc++/signal.h>
 
 #include <map>
 #include <set>
@@ -453,84 +454,6 @@ protected:
 
 	virtual void get_values_vfunc(std::map<Time, ValueBase> &x) const;
 }; // END of class LinkableValueNode
-
-class ValueNodeRegistry {
-public:
-	//! Type that represents a pointer to a ValueNode's constructor
-	using Factory = LinkableValueNode* (*)(const ValueBase&);
-	//! Pointer to check_type method
-	using CheckType = bool (*)(Type &type);
-
-	struct BookEntry
-	{
-		String local_name;
-		Factory factory;
-		CheckType check_type;
-		ReleaseVersion release_version; // which version of synfig introduced this valuenode type
-	};
-
-	using Book = std::map<String,BookEntry>;
-
-public:
-	static Book& book()
-	{
-		if (!book_)
-			book_.reset(new Book());
-		return *book_;
-	}
-
-	static void register_node_type(String name, String local_name, ReleaseVersion version, Factory factory, CheckType check_type)
-	{
-		book()[name].local_name = local_name;
-		book()[name].release_version = version;
-		book()[name].factory = factory;
-		book()[name].check_type = check_type;
-	}
-
-	static bool cleanup()
-	{
-		book_.reset();
-		return true;
-	}
-
-public:
-	//! Creates a Linkable Value Node based on the name and the returned
-	//! value type. Returns a valid Handle if both (name and type) match
-	static LinkableValueNode::Handle create(const String &name, const ValueBase& x, etl::loose_handle<Canvas> canvas);
-	//! Each derived Linkable Value Node has to implement this fucntion and
-	//! should return true only if the type matches. \name is the name of
-	//! the linked value node and \x is the returned value type
-	static bool check_type(const String &name, Type &x);
-
-private:
-	static std::unique_ptr<Book> book_;
-}; // END of class ValueNodeRegistry
-
-// Automatically register class
-// See http://stackoverflow.com/questions/401621/best-way-to-for-c-types-to-self-register-in-a-list
-template<class T>
-class Registrable {
-private:
-	struct do_register {
-		do_register() {
-			T t;
-			ValueNodeRegistry::register_node_type(
-				t.get_name(),
-				t.get_local_name(),
-				T::release_version,
-				reinterpret_cast<ValueNodeRegistry::Factory>(&T::create),
-				&T::check_type
-			);
-		}
-	};
-	// template magic to make sure do_register is instantiated
-	template<do_register&> struct register_ref {};
-	static do_register register_obj;
-	static register_ref<register_obj> ref_obj;
-}; // END of class Registrable
-
-template<class T>
-typename Registrable<T>::do_register Registrable<T>::register_obj;
 
 /*!	\class ValueNodeList
 **	\brief A searchable value_node list container
